+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sales-planner/react
 
+## 0.5.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @sales-planner/shared@0.20.5
+  - @sales-planner/http-client@0.22.5
+
 ## 0.5.4
 
 ### Patch Changes